feat: Added openapi generation cli and renamed openapi-check crate to openapi-cli
