+++ .husky/pre-commit
@@ -1,5 +1,5 @@
 #!/usr/bin/env sh
 
-pnpm backend:openapi-check
+pnpm backend:openapi-cli
 pnpm test
 pnpm lint
