+++ package.json
@@ -17,7 +17,7 @@
     "backend:dev": "pnpm --prefix ./backend/ dev",
     "backend:lint": "pnpm --prefix ./backend/ lint",
     "backend:lint:fix": "pnpm --prefix ./backend/ lint:fix",
-    "backend:openapi-check": "pnpm --prefix ./backend/ openapi-check",
+    "backend:openapi-cli": "pnpm --prefix ./backend/ openapi-cli",
     "test": "pnpm run /:test$/",
     "dev": "pnpm run /:dev$/",
     "lint": "pnpm run /:lint$/",
